+++ src/main/kotlin/kondorcet/Ballot.kt
@@ -1,7 +1,5 @@
 package kondorcet
 
-import java.util.*
-
 /**
  * Represent a ballot
  */
@@ -12,6 +10,10 @@ interface Ballot<out T : Any> {
      */
     val orderedCandidates: List<Set<T>>
 
+    /** Candidates */
+    val candidates: Set<T>
+        get() = orderedCandidates.flatten().toSet()
+
     /**
      * List of winner (many candidates mean ex aequo)
      */
@@ -19,27 +21,13 @@ interface Ballot<out T : Any> {
         get() = orderedCandidates.firstOrNull() ?: emptySet()
 
     /**
-     * Single winner (if ther is one)
+     * Single winner (if there is one)
      */
     val winner: T?
         get() = winners.let { if (it.size == 1) it.first() else null }
 
     /**
-     * Set containing all the candidates
-     */
-    fun candidates(): Set<T> =
-            orderedCandidates.fold(emptySet<T>()) { set, elt -> set + elt }
-
-    /**
      * Return true if, and only if, the ballot contains at least one candidate twice
      */
-    fun hasDuplicates(): Boolean {
-        val set = HashSet<T>()
-
-        for (stage in orderedCandidates)
-            for (candidate in stage)
-                if (!set.add(candidate)) return true
-
-        return false
-    }
+    fun hasDuplicates() = candidates.size != orderedCandidates.flatten().size
 }