+++ src/main/kotlin/kondorcet/method/SchulzeMethod.kt
@@ -1,10 +1,10 @@
 package kondorcet.method
 
-import kondorcet.method.SchulzeMethod.result
+import kondorcet.graph.*
+import kondorcet.utils.max
 import kondorcet.method.SchulzeMethod.resultOf
-import kraft.DirectedPseudoGraph
-import kraft.DirectedWeightedPseudoGraph
-import kraft.algo.widestPaths
+import kondorcet.utils.min
+import kondorcet.utils.orZero
 
 /**
  * Schulze method poll. It ensure to get a single winner and this winner will be a condorcet winner if possible.
@@ -15,6 +15,49 @@ import kraft.algo.widestPaths
  */
 object SchulzeMethod : GraphBasedMethod() {
 
-    override fun <T : Any> result(graph: DirectedWeightedPseudoGraph<T, Int>): DirectedPseudoGraph<T> =
-            graph.widestPaths(0).apply { simplify() }
+    override fun <T : Any> ballotOf(graph: Graph<T, Int>) =
+            CondorcetMethod.ballotOf(graph.widestPath().simplify())
+
+    /**
+     * Returns the widest paths of the graph with the Floyd Warshall Algorithm
+     */
+    fun <T : Any> Graph<T, Int>.widestPath(): Graph<T, Int> {
+        var result: Graph<T, Int> = SimpleGraph(vertices)
+
+        for ((source, target, weight) in edges) {
+            val i = source
+            val j = target
+
+            if (i != j) {
+                val ij = weight
+                val ji = result[j, i].orZero()
+
+                if (ij > ji)
+                    result += Edge(i, j, ij)
+                else
+                    result -= i to j
+            }
+        }
+
+        for (i in vertices) {
+            for (j in (vertices - i)) {
+                for (k in (vertices - i - j)) {
+                    val weight = max(
+                            result[j, k].orZero(),
+                            min(
+                                    result[j, i].orZero(),
+                                    result[i, k].orZero()
+                            )
+                    )
+
+                    if (weight > 0)
+                        result += Edge(j, k, weight)
+                    else
+                        result -= j to k
+                }
+            }
+        }
+
+        return result
+    }
 }